+++ ios/Classes/lib/SwiftAwesomeNotificationsPlugin.swift
@@ -1,17 +1,24 @@
-
-#if !ACTION_EXTENSION
 import UIKit
 import Flutter
 import UserNotifications
 import IosAwnCore
 
-public class SwiftAwesomeNotificationsPlugin:
+protocol AttachablePlugin {
+    func attachPlugin(with registrar: FlutterPluginRegistrar)
+}
+
+public class BaseAwesomeNotificationsPlugin:
                 NSObject,
                 FlutterPlugin,
+                AttachablePlugin,
                 AwesomeEventListener,
                 UNUserNotificationCenterDelegate
 {
     static let TAG = "AwesomeNotificationsPlugin"
+    
+    func attachPlugin(with registrar: FlutterPluginRegistrar) {
+        fatalError("Must Override")
+    }
     
     static var flutterRegistrantCallback: FlutterPluginRegistrantCallback?
     var awesomeNotifications:AwesomeNotifications?
@@ -26,7 +33,7 @@ public class SwiftAwesomeNotificationsPlugin:
             name: Definitions.CHANNEL_FLUTTER_PLUGIN,
             binaryMessenger: registrar.messenger())
         
-        SwiftAwesomeNotificationsPlugin()
+        BaseAwesomeNotificationsPlugin()
             .AttachAwesomeNotificationsPlugin(
                 usingRegistrar: registrar,
                 throughFlutterChannel: flutterChannel)
@@ -57,21 +64,21 @@ public class SwiftAwesomeNotificationsPlugin:
             registrar.addMethodCallDelegate(self, channel: self.flutterChannel!)
             registrar.addApplicationDelegate(self)
             
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, "Awesome Notifications plugin attached to iOS \(floor(NSFoundationVersionNumber))")
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, "Awesome Notifications - App Group : \(Definitions.USER_DEFAULT_TAG)")
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, "Awesome Notifications plugin attached to iOS \(floor(NSFoundationVersionNumber))")
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, "Awesome Notifications - App Group : \(Definitions.USER_DEFAULT_TAG)")
             
             if !Definitions.USER_DEFAULT_TAG.starts(with: "group.") {
                 throw ExceptionFactory
                         .shared
                         .createNewAwesomeException(
-                            className: SwiftAwesomeNotificationsPlugin.TAG,
+                            className: BaseAwesomeNotificationsPlugin.TAG,
                             code: ExceptionCode.CODE_INITIALIZATION_EXCEPTION,
                             message: "Your App Group name \"\(Definitions.USER_DEFAULT_TAG)\" is invalid. It must starts with \"group.\"",
                             detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS+".customAppGroup")
             }
         }
         catch {
-            Logger.shared.e(SwiftAwesomeNotificationsPlugin.TAG, error.localizedDescription)
+            Logger.shared.e(BaseAwesomeNotificationsPlugin.TAG, error.localizedDescription)
         }
     }
     
@@ -85,7 +92,7 @@ public class SwiftAwesomeNotificationsPlugin:
         awesomeNotifications = nil
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, "Awesome Notifications plugin detached from iOS \(floor(NSFoundationVersionNumber))")
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, "Awesome Notifications plugin detached from iOS \(floor(NSFoundationVersionNumber))")
         }
     }
     
@@ -107,7 +114,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 = ExceptionFactory
                         .shared
                         .createNewAwesomeException(
-                            className: SwiftAwesomeNotificationsPlugin.TAG,
+                            className: BaseAwesomeNotificationsPlugin.TAG,
                             code: ExceptionCode.CODE_INITIALIZATION_EXCEPTION,
                             message: "Awesome notifications is currently not available",
                             detailedCode: ExceptionCode.DETAILED_INITIALIZATION_FAILED+".awesomeNotifications.core")
@@ -273,7 +280,7 @@ public class SwiftAwesomeNotificationsPlugin:
                     throw ExceptionFactory
                         .shared
                         .createNewAwesomeException(
-                            className: SwiftAwesomeNotificationsPlugin.TAG,
+                            className: BaseAwesomeNotificationsPlugin.TAG,
                             code: ExceptionCode.CODE_MISSING_METHOD,
                             message: "method \(call.method) not found",
                             detailedCode: ExceptionCode.DETAILED_MISSING_METHOD+"."+call.method)
@@ -292,7 +299,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_UNKNOWN_EXCEPTION,
                         detailedCode: ExceptionCode.DETAILED_UNEXPECTED_ERROR,
                         originalException: error)
@@ -329,7 +336,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Channel data is invalid",
                         detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".channel.data")
@@ -339,7 +346,7 @@ public class SwiftAwesomeNotificationsPlugin:
                             .setChannel(channel: channel) ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.e(SwiftAwesomeNotificationsPlugin.TAG, "Channel \(updated ? "" : "wasn't ")updated")
+            Logger.shared.e(BaseAwesomeNotificationsPlugin.TAG, "Channel \(updated ? "" : "wasn't ")updated")
         }
 		
 		result(updated)
@@ -352,7 +359,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Empty channel key",
                         detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".channel.key")
@@ -362,13 +369,13 @@ public class SwiftAwesomeNotificationsPlugin:
             .removeChannel(channelKey: channelKey) ?? false {
             
             if AwesomeNotifications.debug {
-                Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, "Channel removed")
+                Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, "Channel removed")
             }
             result(true)
         }
         else {
             if AwesomeNotifications.debug {
-                Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, "Channel '\(channelKey)' not found")
+                Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, "Channel '\(channelKey)' not found")
             }
             result(false)
         }
@@ -397,7 +404,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Invalid Badge value",
                         detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".badge.value")
@@ -448,7 +455,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                 .shared
                 .createNewAwesomeException(
-                    className: SwiftAwesomeNotificationsPlugin.TAG,
+                    className: BaseAwesomeNotificationsPlugin.TAG,
                     code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                     message: "Invalid id value",
                     detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".dismiss.id")
@@ -459,7 +466,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .dismissNotification(byId: notificationId!) ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, dismissed ?
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, dismissed ?
                   "Notification \(notificationId!) dismissed":
                   "Notification \(notificationId!) was not found")
         }
@@ -473,7 +480,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Invalid id value",
                         detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".dismiss.id")
@@ -484,7 +491,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .cancelSchedule(byId: notificationId!) ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, cancelled ?
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, cancelled ?
                   "Schedule \(notificationId!) cancelled":
                   "Schedule \(notificationId!) was not found")
         }
@@ -498,7 +505,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Invalid id value",
                         detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".dismiss.id")
@@ -509,7 +516,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .cancelNotification(byId: notificationId!) ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, cancelled ?
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, cancelled ?
                   "Notification \(notificationId!) cancelled":
                   "Notification \(notificationId!) was not found")
         }
@@ -522,7 +529,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Invalid channel key value",
                         detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".dismiss.channelKey")
@@ -533,7 +540,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .dismissNotifications(byChannelKey: channelKey) ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, success ?
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, success ?
                   "Notifications from channel \(channelKey) dismissed":
                   "Notifications from channel \(channelKey) not found")
         }
@@ -546,7 +553,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Invalid channel key value",
                         detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".dismiss.channelKey")
@@ -557,7 +564,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .cancelSchedules(byChannelKey: channelKey) ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, success ?
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, success ?
                   "Scheduled notifications from channel \(channelKey) canceled":
                   "Scheduled notifications from channel \(channelKey) not found")
         }
@@ -570,7 +577,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Invalid channel key value",
                         detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".dismiss.channelKey")
@@ -581,7 +588,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .cancelNotifications(byChannelKey: channelKey) ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, success ?
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, success ?
                   "Notifications and schedules from channel \(channelKey) canceled":
                   "Notifications and schedules from channel \(channelKey) not found")
         }
@@ -594,7 +601,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Invalid group key value",
                         detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".dismiss.groupKey")
@@ -605,7 +612,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .dismissNotifications(byGroupKey: groupKey) ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, success ?
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, success ?
                   "Notifications from group \(groupKey) dismissed":
                   "Notifications from group \(groupKey) not found")
         }
@@ -618,7 +625,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Invalid group key value",
                         detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".dismiss.groupKey")
@@ -629,7 +636,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .cancelSchedules(byGroupKey: groupKey) ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, success ?
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, success ?
                   "Scheduled notifications from group \(groupKey) cancelled":
                   "Scheduled notifications from group \(groupKey) not found")
         }
@@ -642,7 +649,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Invalid group key value",
                         detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".dismiss.groupKey")
@@ -653,7 +660,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .cancelNotifications(byGroupKey: groupKey) ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, success ?
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, success ?
                   "Notifications and schedules from group \(groupKey) cancelled":
                   "Notifications and schedules from group \(groupKey) not found")
         }
@@ -667,7 +674,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .dismissAllNotifications() ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, "All notifications was dismissed")
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, "All notifications was dismissed")
         }
         
         result(success)
@@ -679,7 +686,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .cancelAllSchedules() ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, "All schedules was cancelled")
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, "All schedules was cancelled")
         }
         
         result(success)
@@ -691,7 +698,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 .cancelAllNotifications() ?? false
         
         if AwesomeNotifications.debug {
-            Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, "All notifications was cancelled")
+            Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, "All notifications was cancelled")
         }
         
         result(success)
@@ -798,7 +805,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_MISSING_ARGUMENTS,
                         message: "Arguments are missing",
                         detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS)
@@ -809,7 +816,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Permission list is required",
                         detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS+".permissionList")
@@ -819,7 +826,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Permission list is required",
                         detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS+".permissionList")
@@ -840,7 +847,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_MISSING_ARGUMENTS,
                         message: "Arguments are missing",
                         detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS)
@@ -851,7 +858,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Permission list is required",
                         detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS+".permissionList")
@@ -861,7 +868,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Permission list is required",
                         detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS+".permissionList")
@@ -882,7 +889,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_MISSING_ARGUMENTS,
                         message: "Arguments are missing",
                         detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS)
@@ -893,7 +900,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Permission list is required",
                         detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS+".permissionList")
@@ -903,7 +910,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Permission list is required",
                         detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS+".permissionList")
@@ -925,7 +932,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                         message: "Notification content is invalid",
                         detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS+".notificationModel.data")
@@ -949,7 +956,7 @@ public class SwiftAwesomeNotificationsPlugin:
                             let awesomeException = ExceptionFactory
                                 .shared
                                 .createNewAwesomeException(
-                                    className: SwiftAwesomeNotificationsPlugin.TAG,
+                                    className: BaseAwesomeNotificationsPlugin.TAG,
                                     code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                                     message: "Notification content is invalid",
                                     detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS+".notificationModel.data")
@@ -977,7 +984,7 @@ public class SwiftAwesomeNotificationsPlugin:
             throw ExceptionFactory
                     .shared
                     .createNewAwesomeException(
-                        className: SwiftAwesomeNotificationsPlugin.TAG,
+                        className: BaseAwesomeNotificationsPlugin.TAG,
                         code: ExceptionCode.CODE_MISSING_ARGUMENTS,
                         message: "Arguments are missing",
                         detailedCode: ExceptionCode.DETAILED_REQUIRED_ARGUMENTS)
@@ -996,7 +1003,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 throw ExceptionFactory
                         .shared
                         .createNewAwesomeException(
-                            className: SwiftAwesomeNotificationsPlugin.TAG,
+                            className: BaseAwesomeNotificationsPlugin.TAG,
                             code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                             message: "Notification channel `\(channelsData)` is invalid",
                             detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".channel.invalid.\(channelsData)")
@@ -1007,7 +1014,7 @@ public class SwiftAwesomeNotificationsPlugin:
                 throw ExceptionFactory
                         .shared
                         .createNewAwesomeException(
-                            className: SwiftAwesomeNotificationsPlugin.TAG,
+                            className: BaseAwesomeNotificationsPlugin.TAG,
                             code: ExceptionCode.CODE_INVALID_ARGUMENTS,
                             message: "Notification channel `\(channelsData)` is invalid",
                             detailedCode: ExceptionCode.DETAILED_INVALID_ARGUMENTS+".channel.invalid.\(channelsData)")
@@ -1023,7 +1030,7 @@ public class SwiftAwesomeNotificationsPlugin:
                     backgroundHandle: dartBgHandle,
                     debug: debug)
 		
-		Logger.shared.d(SwiftAwesomeNotificationsPlugin.TAG, "Awesome Notifications service initialized")
+		Logger.shared.d(BaseAwesomeNotificationsPlugin.TAG, "Awesome Notifications service initialized")
 		result(awesomeNotifications != nil)
     }
     
@@ -1045,10 +1052,23 @@ public class SwiftAwesomeNotificationsPlugin:
         
         let success = actionHandle != 0
         if !success {
-            Logger.shared.e(SwiftAwesomeNotificationsPlugin.TAG, "Attention: there is no valid static method to receive notification action data in background")
+            Logger.shared.e(BaseAwesomeNotificationsPlugin.TAG, "Attention: there is no valid static method to receive notification action data in background")
         }
         
         result(success)
     }
 }
-#endif
+
+public class SwiftAwesomeNotificationsPlugin: BaseAwesomeNotificationsPlugin {
+    public override func attachPlugin(with registrar: FlutterPluginRegistrar) {
+        registrar.addMethodCallDelegate(self, channel: self.flutterChannel!)
+    }
+}
+
+@available(iOSApplicationExtension, unavailable)
+public class ExtensionAwesomeNotificationsPlugin: BaseAwesomeNotificationsPlugin {
+    public override func attachPlugin(with registrar: FlutterPluginRegistrar) {
+        registrar.addMethodCallDelegate(self, channel: self.flutterChannel!)
+        registrar.addApplicationDelegate(self)
+    }
+}