+++ lib/src/models/notification_content.dart
@@ -1,9 +1,6 @@
-import 'package:flutter/material.dart';
-
 import '../definitions.dart';
-import '../enumerators/action_type.dart';
-import '../enumerators/notification_category.dart';
 import '../enumerators/notification_layout.dart';
+import '../enumerators/notification_play_state.dart';
 import '../utils/assert_utils.dart';
 import 'base_notification_content.dart';
 
@@ -11,7 +8,9 @@ import 'base_notification_content.dart';
 /// If notification has no [body] or [title], it will only be created, but not displayed to the user (background notification).
 class NotificationContent extends BaseNotificationContent {
   bool? _hideLargeIconOnExpand;
-  int? _progress, _badge, _duration, _playState;
+  int? _progress, _badge;
+  Duration? _duration;
+  NotificationPlayState? _playState;
   String? _ticker;
   double? _playbackSpeed;
 
@@ -54,11 +53,11 @@ class NotificationContent extends BaseNotificationContent {
     return _locked;
   }
 
-  int? get duration {
+  Duration? get duration {
     return _duration;
   }
 
-  int? get playState {
+  NotificationPlayState? get playState {
     return _playState;
   }
 
@@ -67,31 +66,31 @@ class NotificationContent extends BaseNotificationContent {
   }
 
   NotificationContent(
-      {required int id,
-      required String channelKey,
-      String? title,
-      String? body,
-      String? groupKey,
-      String? summary,
-      String? icon,
-      String? largeIcon,
-      String? bigPicture,
-      String? customSound,
-      bool showWhen = true,
-      bool wakeUpScreen = false,
-      bool fullScreenIntent = false,
-      bool criticalAlert = false,
-      bool roundedLargeIcon = false,
-      bool roundedBigPicture = false,
-      bool autoDismissible = true,
-      Color? color,
-      Duration? timeoutAfter,
-      Duration? chronometer,
-      Color? backgroundColor,
-      ActionType actionType = ActionType.Default,
+      {required int super.id,
+      required String super.channelKey,
+      super.title,
+      super.body,
+      super.groupKey,
+      super.summary,
+      super.icon,
+      super.largeIcon,
+      super.bigPicture,
+      super.customSound,
+      super.showWhen,
+      super.wakeUpScreen,
+      super.fullScreenIntent,
+      super.criticalAlert,
+      super.roundedLargeIcon,
+      super.roundedBigPicture,
+      super.autoDismissible,
+      super.color,
+      super.timeoutAfter,
+      super.chronometer,
+      super.backgroundColor,
+      super.actionType,
       NotificationLayout notificationLayout = NotificationLayout.Default,
-      Map<String, String?>? payload,
-      NotificationCategory? category,
+      super.payload,
+      super.category,
       bool hideLargeIconOnExpand = false,
       bool locked = false,
       int? progress,
@@ -99,8 +98,8 @@ class NotificationContent extends BaseNotificationContent {
       String? ticker,
       bool displayOnForeground = true,
       bool displayOnBackground = true,
-      int? duration,
-      int? playState,
+      Duration? duration,
+      NotificationPlayState? playState,
       double? playbackSpeed})
       : _hideLargeIconOnExpand = hideLargeIconOnExpand,
         _progress = progress,
@@ -112,32 +111,7 @@ class NotificationContent extends BaseNotificationContent {
         _locked = locked,
         _duration = duration,
         _playState = playState,
-        _playbackSpeed = playbackSpeed,
-        super(
-            id: id,
-            channelKey: channelKey,
-            groupKey: groupKey,
-            title: title,
-            body: body,
-            summary: summary,
-            wakeUpScreen: wakeUpScreen,
-            fullScreenIntent: fullScreenIntent,
-            category: category,
-            criticalAlert: criticalAlert,
-            showWhen: showWhen,
-            payload: payload,
-            icon: icon,
-            timeoutAfter: timeoutAfter,
-            chronometer: chronometer,
-            largeIcon: largeIcon,
-            bigPicture: bigPicture,
-            customSound: customSound,
-            autoDismissible: autoDismissible,
-            color: color,
-            actionType: actionType,
-            backgroundColor: backgroundColor,
-            roundedBigPicture: roundedBigPicture,
-            roundedLargeIcon: roundedLargeIcon);
+        _playbackSpeed = playbackSpeed;
 
   @override
   NotificationContent? fromMap(Map<String, dynamic> mapData) {
@@ -153,9 +127,10 @@ class NotificationContent extends BaseNotificationContent {
     _locked =
         AwesomeAssertUtils.extractValue<bool>(NOTIFICATION_LOCKED, mapData);
     _duration =
-        AwesomeAssertUtils.extractValue<int>(NOTIFICATION_DURATION, mapData);
+        AwesomeAssertUtils.extractValue<Duration>(NOTIFICATION_DURATION, mapData);
     _playState =
-        AwesomeAssertUtils.extractValue<int>(NOTIFICATION_PLAY_STATE, mapData);
+        NotificationPlayState.fromMap(mapData[NOTIFICATION_PLAY_STATE]);
+
     _playbackSpeed =
         AwesomeAssertUtils.extractValue<double>(NOTIFICATION_PLAYBACK_SPEED, mapData);
 
@@ -191,8 +166,8 @@ class NotificationContent extends BaseNotificationContent {
         NOTIFICATION_LAYOUT: _notificationLayout?.name,
         NOTIFICATION_DISPLAY_ON_FOREGROUND: _displayOnForeground,
         NOTIFICATION_DISPLAY_ON_BACKGROUND: _displayOnBackground,
-        NOTIFICATION_DURATION: _duration,
-        NOTIFICATION_PLAY_STATE: _playState,
+        NOTIFICATION_DURATION: _duration?.inSeconds,
+        NOTIFICATION_PLAY_STATE: _playState?.toMap(),
         NOTIFICATION_PLAYBACK_SPEED: _playbackSpeed,
       });
     return dataMap;